+++ handler/http_route.go
@@ -1,0 +1,11 @@
+package handler
+
+import "net/http"
+
+type HttpRouter struct {
+  RoutePath        string
+  Handler          func(http.ResponseWriter, *http.Request)
+}
+
+
+