+++ CHANGELOG.md
@@ -4,8 +4,11 @@
 - [NEW] support rpc, server & client 
 - [NEW] handler plugins, can use different handler with plugin easy
 - [OPTIMIZE] optimize some code archieture
-- [NEW] sync or async call handler 
 - [NEW] support protobuf data protol 
+- [ ] sync or async call handler 
+- [x] plugin adapter for HTTP frontend 
+- [ ] plugin adapter for HTTP backend 
+- [ ] plugin adapter for RPC 
 
 
 ### MonRabit 0.1